backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
pydantic-settings==2.7.0
python-multipart==0.0.20

joblib==1.4.2
numpy==2.2.1
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.6.0
statsmodels==0.14.4
torch==2.5.1

pymongo==4.10.1
pyzmq==26.2.0
river==0.22.0
